+++ src/game/SceneObjects/GameObject.h
@@ -26,7 +26,7 @@ class GameObject
   CollisionComponent* collisionComponent();
 
   void updateCollisionComponent();
-  void move(float delta_time, float x_dir, float y_dir, float speed);
+  void move(double delta_time, float x_dir, float y_dir, float speed);
   std::vector<float>
   getDirectionFromTo(float from_x, float from_y, float to_x, float to_y);
   float getDistanceBetween(float from_x, float from_y, float to_x, float to_y);
